+++ R/bilateral.R
@@ -168,6 +168,39 @@ geomPaasche_t <- function(p0, p1, q0, q1){
   return(prod((p1/p0)^s1))
 }
 
+#' time product dummy
+#'
+#' @keywords internal
+#' @noRd
+tpd_t <- function(p0, p1, q0, q1){
+
+  exp0 <- sum(p0*q0)
+  exp1 <- sum(p1*q1)
+  s0 <- (p0*q0)/exp0
+  s1 <- (p1*q1)/exp1
+
+  df1 <- data.frame(lnP = log(p1),
+                    D = 1,
+                    s = s1,
+                    product = as.factor(seq_along(p1)))
+
+  df0 <- data.frame(lnP = log(p0),
+                    D = 0,
+                    s = s0,
+                    product = as.factor(seq_along(p0)))
+
+  regData <- rbind(df0, df1)
+
+  reg <- lm(lnP ~ D + product, data = regData, weights = s)
+  coeffs <- coef(reg)
+  vars <- diag(vcov(reg))
+
+  b <- coeffs[which(names(coeffs) == "D")]
+  varB <- vars[which(names(vars) == "D")]
+
+  return(exp(b - 0.5*varB))
+}
+
 #' Computes a bilateral price index
 #'
 #' A function to compute a price index given data on products over time
@@ -182,7 +215,7 @@ geomPaasche_t <- function(p0, p1, q0, q1){
 #' There may be observations on multiple products for each time period.
 #' @param indexMethod A character string to select the index number method. Valid index
 #' number methods are dutot, carli, jevons, laspeyres, paasche, fisher, cswd,
-#' harmonic, tornqvist, satovartia, walsh, CES, geomLaspeyres and geomPaasche.
+#' harmonic, tornqvist, satovartia, walsh, CES, geomLaspeyres, geomPaasche and tpd.
 #' @param sample A character string specifying whether a matched sample
 #' should be used.
 #' @param output A character string specifying whether a chained (output="chained")
@@ -221,7 +254,7 @@ priceIndex <- function(x, pvar, qvar, pervar, indexMethod = "laspeyres", prodID,
   # check that a valid method is chosen
   validMethods <- c("dutot","carli","jevons","harmonic","cswd","laspeyres",
                     "paasche","fisher","tornqvist","satovartia","walsh","ces",
-                    "geomlaspeyres", "geompaasche")
+                    "geomlaspeyres", "geompaasche", "tpd")
 
   if(!(tolower(indexMethod) %in% validMethods)){
     stop("Not a valid index number method.")
@@ -297,8 +330,11 @@ priceIndex <- function(x, pvar, qvar, pervar, indexMethod = "laspeyres", prodID,
 
     # if matching requested then remove unmatched items
     if(sample=="matched"){
+
+      # remove unmatched products
       xt1 <- xt1[xt1[[prodID]] %in% unique(xt0[[prodID]]),]
       xt0 <- xt0[xt0[[prodID]] %in% unique(xt1[[prodID]]),]
+
     }
 
     # set the price index element to NA if there are no
@@ -329,7 +365,8 @@ priceIndex <- function(x, pvar, qvar, pervar, indexMethod = "laspeyres", prodID,
              walsh = {plist[i,1] <- walsh_t(p0,p1,q0,q1)},
              ces = {plist[i,1] <- lloydMoulton_t0(p0,p1,q0,sigma = sigma)},
              geomlaspeyres = {plist[i,1] <- geomLaspeyres_t(p0, p1, q0, q1)},
-             geompaasche = {plist[i,1] <- geomPaasche_t(p0, p1, q0, q1)}
+             geompaasche = {plist[i,1] <- geomPaasche_t(p0, p1, q0, q1)},
+             tpd = {plist[i,1] <- tpd_t(p0, p1, q0, q1)}
       )
 
       # if similarity chain linking then multiply the index by the link period index